Fix critical review issues CR-001 through CR-004
CR-001: Cast to uint16_t before << 8 in all sensorToHost16 and BMP180
calibration lambdas to prevent UB on 16-bit int platforms.
CR-002: ADXL345 scale factor now computed from range+fullRes config
instead of hardcoded constant. Fixes wrong readings when fullRes=false.
CR-003: Remove dead OTP_WORD_SENS_X constant, document dual-use word.
CR-004: Guard BMP180 computeB5 and computeTruePressure against div-by-zero
with zero-initialized calibration data.
Also: CR-008 BMM350 temp register magic number → named constant.
Also: CR-012 ADXL345 sensitivity corrected from 4.0 to 3.9 mg/LSB.

+++ drivers/lps22df/LPS22DF.cpp
@@ -22,7 +22,8 @@ LPS22DF::LPS22DF(II2CBus& bus, IDelayProvider& delay, const LPS22DFConfig& cfg)
 {}
 
 int16_t LPS22DF::sensorToHost16(const uint8_t* buf) {
-    return static_cast<int16_t>((buf[1] << 8) | buf[0]);
+    return static_cast<int16_t>(
+        (static_cast<uint16_t>(buf[1]) << 8) | static_cast<uint16_t>(buf[0]));
 }
 
 int32_t LPS22DF::sensorToHost24(const uint8_t* buf) {
